requirements.txt
absl-py==0.12.0
astunparse==1.6.3
attrs==20.3.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dill==0.3.3
Flask==1.1.2
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==1.28.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.34.1
h5py==3.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==1.1.1
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
Pillow==8.1.1
promise==2.3
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0rc0
tensorflow-datasets==4.2.0
tensorflow-metadata==0.29.0
termcolor==1.1.0
tf-estimator-nightly==2.5.0.dev2021032501
torch==1.8.0
tqdm==4.59.0
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1
